Move ILoggerListener
