+++ src/SLCore.Listeners.UnitTests/Logging/LoggerSerializationTests.cs
@@ -19,6 +19,7 @@
  */
 
 using Newtonsoft.Json;
+using SonarLint.VisualStudio.SLCore.Listener.Logger;
 using SonarLint.VisualStudio.SLCore.Listeners.Implementation;
 
 namespace SonarLint.VisualStudio.SLCore.Listeners.UnitTests.Logging
